Add some podcast urls (#5)
* Add BBC 6 Minute English podcast to the list

This commit adds 'BBC 6 Minute English' podcast to the list. There are
already some BBC resources on the list. But I think this podcast is
good to learn some English words in 6 minutes.

* Add CNN 10 to the podcasts list

This commit adds 'CNN 10' to the podcasts list. We can learn English
through the latest news with this.

* Add Talk Python To Me Podcast

This commit adds 'Talk Python To Me' podcast to the list. We can learn
English and python technology from this podcast. There are many episodes
already and updated constantly. Of course, they're super fun episodes.

* Add 'The CultCast' podcast to the list

This commit adds 'The CultCast' podcast to the list. This podcast is
talking about Apple products and technologies such as Mac, AppleWatch,
macOS, iOS, etc. Their talks are super fast by native English speakers
but super fun.

* Revert "Add Talk Python To Me Podcast"

This reverts commit 89ab918715820d58f754d184482ec374cc6b386c.